% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_special_pop_url}
\alias{build_special_pop_url}
\title{Build URL for special populations file}
\usage{
build_special_pop_url(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
URL string
}
\description{
Build URL for special populations file
}
\keyword{internal}
