% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\docType{data}
\name{DESE_ASSESS_SOCRATA_API}
\alias{DESE_ASSESS_SOCRATA_API}
\title{Base URL for the DESE Assessment Socrata API}
\format{
An object of class \code{character} of length 1.
}
\usage{
DESE_ASSESS_SOCRATA_API
}
\description{
Base URL for the DESE Assessment Socrata API
}
\keyword{internal}
