% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\docType{data}
\name{DESE_GRAD_SOCRATA_API}
\alias{DESE_GRAD_SOCRATA_API}
\title{Base URL for the DESE Graduation Socrata API}
\format{
An object of class \code{character} of length 1.
}
\usage{
DESE_GRAD_SOCRATA_API
}
\description{
Base URL for the DESE Graduation Socrata API
}
\keyword{internal}
