% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_enrollment_file}
\alias{download_enrollment_file}
\title{Download a single enrollment file}
\usage{
download_enrollment_file(end_year, level, type)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"district" or "school"}

\item{type}{"race", "gender", or "grade"}
}
\value{
Data frame with enrollment data
}
\description{
Download a single enrollment file
}
\keyword{internal}
