% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_special_populations}
\alias{download_special_populations}
\title{Download special populations data}
\usage{
download_special_populations(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with special populations data, or NULL if not available
}
\description{
Download special populations data
}
\keyword{internal}
