% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_available_assess_subgroups}
\alias{get_available_assess_subgroups}
\title{Get available assessment subgroups}
\usage{
get_available_assess_subgroups()
}
\value{
Character vector of subgroup names (lowercase, underscore-separated)
}
\description{
Returns a list of available student subgroups in the assessment data.
}
\examples{
\dontrun{
get_available_assess_subgroups()
# Returns: "all", "male", "female", "white", "black", "hispanic", ...
}
}
