% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_available_assess_years}
\alias{get_available_assess_years}
\title{Get available assessment years}
\usage{
get_available_assess_years()
}
\value{
Integer vector of years (2017-2019, 2021-2025)
}
\description{
Returns a vector of years for which MCAS assessment data is available
from the Massachusetts DESE Socrata API.
}
\examples{
\dontrun{
get_available_assess_years()
# Returns: 2017 2018 2019 2021 2022 2023 2024 2025
}
}
