% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Massachusetts enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available school years (end year)
}
\description{
Returns the range of school years for which enrollment data is available
from the DESE Socrata API. Data is available from 1994-2024.
}
\details{
Note: Years 1992-1993 exist in the API but have very limited data
(only special populations). Full enrollment data starts in 1994.
}
\examples{
get_available_years()
# Returns: 1994, 1995, ..., 2023, 2024
}
