% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_types.R
\docType{data}
\name{MASSGIS_SCHOOLS_URL}
\alias{MASSGIS_SCHOOLS_URL}
\title{MassGIS Schools shapefile URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
MASSGIS_SCHOOLS_URL
}
\description{
MassGIS Schools shapefile URL
}
\keyword{internal}
