% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_district_enr}
\alias{process_district_enr}
\title{Process district-level enrollment data}
\usage{
process_district_enr(race_df, special_df, end_year)
}
\arguments{
\item{race_df}{Raw district race data frame}

\item{special_df}{Raw special populations data frame (can be NULL)}

\item{end_year}{School year end}
}
\value{
Processed district data frame
}
\description{
Process district-level enrollment data
}
\keyword{internal}
