% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graduation.R
\name{process_graduation}
\alias{process_graduation}
\title{Process raw graduation data into standard schema}
\usage{
process_graduation(raw_data, end_year)
}
\arguments{
\item{raw_data}{Data frame from get_raw_graduation()}

\item{end_year}{School year end}
}
\value{
Processed data frame with standardized columns
}
\description{
Transforms raw Socrata API data into the standardized schema used by
the package.
}
\keyword{internal}
