% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{standardize_column_names}
\alias{standardize_column_names}
\title{Standardize column names across different year formats}
\usage{
standardize_column_names(col_names)
}
\arguments{
\item{col_names}{Vector of column names}
}
\value{
Vector of standardized column names
}
\description{
Standardize column names across different year formats
}
\keyword{internal}
