% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{standardize_grade}
\alias{standardize_grade}
\title{Standardize grade level strings}
\usage{
standardize_grade(grade)
}
\arguments{
\item{grade}{Vector of grade strings}
}
\value{
Vector of standardized grade codes
}
\description{
Standardize grade level strings
}
\keyword{internal}
