% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yr_code_to_end_year}
\alias{yr_code_to_end_year}
\title{Map school year codes to end year}
\usage{
yr_code_to_end_year(yr_code)
}
\arguments{
\item{yr_code}{School year code string}
}
\value{
Integer end year
}
\description{
Converts DESE school year codes (e.g., "2324") to end year (2024)
}
\keyword{internal}
